# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Octroi API
  description: API gateway for AI agent tool access.
  version: "0.1.0"
servers:
  - url: /
    description: Default server
...
security: []
paths:
...
  /api/v1/tools:
    get:
      operationId: listTools
      tags: [public]
      summary: List tools
      parameters:
        - $ref: "#/components/parameters/QueryQ"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryCursor"
      responses:
        "200":
          description: Paginated list of tools.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Public: Get tool ----------
  /api/v1/tools/{id}:
    get:
      operationId: getTool
      tags: [public]
      summary: Get a tool by ID
      parameters:
        - $ref: "#/components/parameters/PathToolID"
      responses:
        "200":
          description: The tool (public view, no endpoint or auth_config).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolPublic"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Admin: Tool CRUD ----------
  /api/v1/admin/tools:
    post:
      operationId: createTool
      tags: [admin-tools]
      summary: Create a tool
      security:
        - AdminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolInput"
      responses:
        "201":
          description: Tool created (admin view includes endpoint and auth_config).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolAdmin"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/tools/{id}:
    put:
      operationId: updateTool
      tags: [admin-tools]
      summary: Update a tool
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathToolID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateToolInput"
      responses:
        "200":
          description: Tool updated (admin view).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolAdmin"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteTool
      tags: [admin-tools]
      summary: Delete a tool
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathToolID"
      responses:
        "204":
          description: Tool deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Admin: Agent CRUD ----------
...
  /api/v1/admin/usage:
    get:
      operationId: getUsageAdmin
      tags: [admin-usage]
      summary: Get usage summary (admin)
      description: Query usage summary across all or filtered agents and tools.
      security:
        - AdminBearer: []
      parameters:
        - name: agent_id
          in: query
          schema:
            type: string
          description: Filter by agent ID.
        - name: tool_id
          in: query
          schema:
            type: string
          description: Filter by tool ID.
        - $ref: "#/components/parameters/QueryFrom"
        - $ref: "#/components/parameters/QueryTo"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryCursor"
      responses:
        "200":
          description: Usage summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/usage/agents/{agentID}:
    get:
      operationId: getUsageByAgent
      tags: [admin-usage]
      summary: Get usage summary for a specific agent
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathBudgetAgentID"
        - $ref: "#/components/parameters/QueryFrom"
        - $ref: "#/components/parameters/QueryTo"
      responses:
        "200":
          description: Usage summary for the agent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/usage/tools/{toolID}:
    get:
      operationId: getUsageByTool
      tags: [admin-usage]
      summary: Get usage summary for a specific tool
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathBudgetToolID"
        - $ref: "#/components/parameters/QueryFrom"
        - $ref: "#/components/parameters/QueryTo"
      responses:
        "200":
          description: Usage summary for the tool.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/usage/agents/{agentID}/tools/{toolID}:
    get:
      operationId: getUsageByAgentTool
      tags: [admin-usage]
      summary: Get usage summary for a specific agent and tool combination
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathBudgetAgentID"
        - $ref: "#/components/parameters/PathBudgetToolID"
        - $ref: "#/components/parameters/QueryFrom"
        - $ref: "#/components/parameters/QueryTo"
      responses:
        "200":
          description: Usage summary for the agent-tool pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/usage:
    get:
      operationId: getUsage
      tags: [agent]
      summary: Get usage summary for the authenticated agent
      security:
        - AgentBearer: []
      parameters:
        - name: tool_id
          in: query
          schema:
            type: string
          description: Filter by tool ID.
        - $ref: "#/components/parameters/QueryFrom"
        - $ref: "#/components/parameters/QueryTo"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryCursor"
      responses:
        "200":
          description: Usage summary scoped to the authenticated agent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/usage/transactions:
    get:
      operationId: listTransactions
      tags: [agent]
      summary: List transactions for the authenticated agent
      security:
        - AgentBearer: []
      parameters:
        - name: tool_id
          in: query
          schema:
            type: string
          description: Filter by tool ID.
        - $ref: "#/components/parameters/QueryFrom"
        - $ref: "#/components/parameters/QueryTo"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryCursor"
      responses:
        "200":
          description: Paginated list of transactions scoped to the authenticated agent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Proxy ----------
...
components:
...
  securitySchemes:
    AdminBearer:
      type: http
      scheme: bearer
      description: Admin API key passed as a Bearer token.
    AgentBearer:
      type: http
      scheme: bearer
      description: Agent API key passed as a Bearer token.

  # ---------- Reusable parameters ----------
  parameters:
    PathToolID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Tool ID.
...
    PathBudgetAgentID:
      name: agentID
      in: path
      required: true
      schema:
        type: string
      description: Agent ID.
    PathBudgetToolID:
      name: toolID
      in: path
      required: true
      schema:
        type: string
      description: Tool ID.
...
    QueryQ:
      name: q
      in: query
      schema:
        type: string
      description: Free-text search query.
    QueryLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
      description: Maximum number of items to return.
    QueryCursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor for fetching the next page.
    QueryFrom:
      name: from
      in: query
      schema:
        type: string
      description: Start date/time filter (YYYY-MM-DD or RFC 3339).
    QueryTo:
      name: to
      in: query
      schema:
        type: string
      description: End date/time filter (YYYY-MM-DD or RFC 3339).

  # ---------- Reusable responses ----------
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Unauthorized.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: tool not found

    # --- Manifest ---
...
    ToolPublic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        auth_type:
          type: string
        tags:
          type: array
          items:
            type: string
        pricing_model:
          type: string
        pricing_amount:
          type: number
          format: double
        pricing_currency:
          type: string
        rate_limit:
          type: integer
        budget_limit:
          type: number
          format: double
        budget_window:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Tool (admin view, includes endpoint and auth_config) ---
    ToolAdmin:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
        auth_type:
          type: string
        auth_config:
          type: object
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
        pricing_model:
          type: string
        pricing_amount:
          type: number
          format: double
        pricing_currency:
          type: string
        rate_limit:
          type: integer
        budget_limit:
          type: number
          format: double
        budget_window:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Tool inputs ---
    CreateToolInput:
      type: object
      required: [name, description, endpoint, auth_type]
      properties:
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
        auth_type:
          type: string
        auth_config:
          type: object
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
        pricing_model:
          type: string
        pricing_amount:
          type: number
          format: double
        pricing_currency:
          type: string
        rate_limit:
          type: integer
        budget_limit:
          type: number
          format: double
        budget_window:
          type: string
    UpdateToolInput:
      type: object
      description: All fields are optional. Only provided fields are applied.
      properties:
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
        auth_type:
          type: string
        auth_config:
          type: object
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
        pricing_model:
          type: string
        pricing_amount:
          type: number
          format: double
        pricing_currency:
          type: string
        rate_limit:
          type: integer
        budget_limit:
          type: number
          format: double
        budget_window:
          type: string
    ToolListResponse:
      type: object
      required: [tools]
      properties:
        tools:
          type: array
          items:
            $ref: "#/components/schemas/ToolPublic"
        next_cursor:
          type: string
          description: Opaque cursor for the next page. Absent when there are no more results.

    # --- Agent ---
...
    UsageSummary:
      type: object
      properties:
        total_requests:
          type: integer
          format: int64
        total_cost:
          type: number
          format: double
        success_count:
          type: integer
          format: int64
        error_count:
          type: integer
          format: int64
        avg_latency_ms:
          type: number
          format: double
    Transaction:
      type: object
      properties:
        id:
          type: string
        agent_id:
          type: string
        tool_id:
          type: string
        timestamp:
          type: string
          format: date-time
        method:
          type: string
        path:
          type: string
        status_code:
          type: integer
        latency_ms:
          type: integer
          format: int64
        request_size:
          type: integer
          format: int64
        response_size:
          type: integer
          format: int64
        success:
          type: boolean
        cost:
          type: number
          format: double
        error:
          type: string
    TransactionListResponse:
      type: object
      required: [transactions]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        next_cursor:
          type: string
          description: Opaque cursor for the next page. Absent when there are no more results.